requirements.txt
PyQt6>=6.4.0
pyserial>=3.5
numpy>=1.21.0
matplotlib>=3.6.0